Start using the useAsync hook in the Tempo TraceQL configuration page to retrieve the tags and datasource
